Create solution.cpp
The Monty Hall problem
The Monty Hall problem is a famous conundrum in probability which takes the form of a hypothetical game show. The contestant is presented with three doors; behind one is a car and behind each of the other two is a goat. The contestant picks a door and then the gameshow host opens a different door to reveal a goat. The host knows which door conceals the car. The contestant is then invited to switch to the other closed door or stick with their initial choice.

Perhaps counter-intuitively, the best strategy to win is to switch, as demonstrated by the following simulation.

Reference: https://scipython.com/book/chapter-4-the-core-python-language-ii/examples/the-monty-hall-problem/